dm-qcow2: Do not fail .ctr in case dirty bit is set
The protocol is (#PSBM-128992):

1)Userspace sets DIRTY bit in top delta.

(https://github.com/qemu/qemu/blob/master/docs/interop/qcow2.txt, incompatible_features: dirty bit).

2)Kernel checks that dirty bit is set.

In case of power down or crash, userspace on mount sees that dirty bit is already set
and it calls "qemu-img check -r all file.qcow2".

Note. Q: Why isn't dirty bit set by kernel? A: We have such the mode in driver,
but it is disabled by default. The reason is that "dmsetup reload" creates
new target before old target is destructed, so synchronization of this is not possible.

Currenly, this protocol is broken in kernel: kernel fails, if it sees dirty bit.
This patch fixes the problem.

Signed-off-by: Kirill Tkhai <ktkhai@virtuozzo.com>